backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.13.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# File handling and OCR
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1

# HTTP requests (for external APIs)
httpx==0.25.2
requests==2.31.0

# Background tasks and async
celery==5.3.4
redis==5.0.1

# AI/ML (실제 AI 서비스 연동)
openai==1.3.7
anthropic==0.7.7
google-generativeai==0.3.2
langchain==0.0.340

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Email (for notifications)
fastapi-mail==1.4.1

# CORS
python-cors==1.0.0

# Utils
python-slugify==8.0.1
validators==0.22.0